Enhanced support for JOIN and PRIVMSG
It was impossible to send to a channel you are not connected to if the channel has mode +n
You can now specify multiple channels to connect to.
Channel passwords can be specified for where mode +k is set.

PRIVMSG can be sent to specific channel by specifing #
eg: echo "#channel message to send"
PRIVMSG can be sent to a user by specifing @
eg: echo "@user message to send"

There is now a defualt option so you can specify which channel PRIVMSG is sent to  when channel / user is not specified.
If this is not set then the message will be broadcast to all connected channels.