Bug fixes, Support for multiline and error checks
Added support for sending multiline messages to bot. It will now send a privmsg per line.
Changed if channels to elif channels, should have all been one if block.
Broadcasting to all channels was broken, was looking for default=False should have been default=None
Remove fork, unecessary.
Checking that when default channel is supplied it is in the list of channels to connect to, otherwise default will not work on when the +n flag is set.